chapter-0/requirements.txt
aws_cdk.core
aws_cdk.aws_sagemaker
aws_cdk.aws_iam
aws_cdk.aws_ec2

